Autopilot LindiPlane: Add options for unsteady flap lift
The unsteady flap aerodynamics can now be boosted or
eliminated. For the controller reference, the delay
due to unsteady flap is now considered, if unsteady
flap aerodynamics are enabled. In lindiPlaneAutoCreate
the controller gains are reduced if unsteady flap
aerodynamics are enabled and not boosted.

It was observed that the pitch axis oscillates
stronger when moving the center of gravity backward
in case direct lift control is enabled and unsteady
flap aerodynamics are enabled. The problem can be
removed if the delayed downwash is not considered
in the direct lift control allocation loop. This
should be investigated further in the future.