Simple wing: Fix lift curve for non-integer data points
When initializing the simple wing, the data points
for lift and drag coefficient stored in the parameter
file are interpolated over the angle of attack.
If non-integer angles of attack are used, errors
occured in the interpolation of the lift curve, as
the angles of attack were subsequently manipulated
when the function simpleWingSetDragCurve was called.
The problem has been solved by interpolating the
interpolated curves again to equidistant angles of
attack in 1-degree steps at the end of both
functions.