Autopilot LindiCopter: Fix ArduCopter firmware crash
With new installations of Ubuntu and/or ArduPilot,
the firmware in the SITL crashed as soon as the
LindiCopter flight mode (mode 29) was activated.
It turned out that the error was caused during
the calculation of the second derivative of the
(normalized) lean vector. In this function, some
division by the norm of the lean vector were
carried out. Apparently, it was divided by zero,
even though the norm should never actually be
zero. Using the divideFinite function solved the
problem.