ArduPilot parameters: Support Simulink Scopes (pointers)
In case of Scopes, there are pointers in the generated
C++ code. The variable names of these pointers were
not detected by the code parser since the * at the
beginning of the variable name was not considered as
start of a variable name. As a consequence, the value
output of the unpackStruct function (m-utils)
was empty and the lengths of the fieldpaths and the
values arrays was different. This caused wrong detection
of the top-level struct of the tunable parameters
struct leading to faulty generated C++ code by the
apPar framework.
The problem was solved by adjusting the function
apPar_getStructsFromHeaderFile so that pointers are
detected. Moreover, to reduce the risk of future
issues, the unpackStruct function was changed to
not support empty variables anymore.