Autopilot LindiPlane: Use new discrete T2 filter
The new T2 filter is now used for low-pass filtering
of the sensors. It assures stability and higher
accuracy at high cutoff frequencies.
The servo dynamics are still modeled with the old
PT2 filters (ode1) because a booster (high-pass filter)
used for the servo dynamics which must be updated as
soon as the new T2 filter is used.
Moreover, the slow low-pass filters (reference modes)
are also sill modeled with the old PT2 filters for
simplicity.