Autopilot LindiPlane: Improve maneuver load alleviation
The desired u_d from the maneuver load alleviation block was
specified as Delta_u_d in the control allocation block. It was
discovered that this led to oscillations in the step response.
The problem was solved by adding u_d directly to the the control
input (after control allocation) and by adding u_d to u and
Delta_u_d in the control allocation block.