Autopilot LindiPlane: Remove notch filters for vertical acceleration (DLC)
LindiPlane runs at 400 Hz by default. This may be not fast enough
for running notch filters on the propeller frequency. Aliasing
was observed already for 75 Hz oscillations (depending on the
filter parameters). Notch filters should therefore run in a
faster loop, as implemented in ArduPilot. The ArduPilot notch
filters run in the inertial sensor backend at 2 kHz by default
and can be increased up to 8 kHz.